Remove the manifest section
The idea of this PR is that the "core" Web Bundles specification does not
include the manifest sections, but that these sections can be defined in other
documents. The motivation is to factor the Web Bundles specification to include
just the minimal amount needed to support cases like subresource loading.
Subresource loading does not need a manifest.

Adapted from https://github.com/WICG/webpackage/pull/619